* Structural change to x86/Evergreen simulation loops (emulation and timing   simulation to allow for concurrent execution of CPU and GPU, affecting many   files. * New mechanism to suspend x86 contexts, based on callback functions checking   for context wake up. * Non-blocking implementation of 'clFinish' in OpenCL API. The x86 context   suspends, but the main simulation loop continues until a command queue   empties.
git-svn-id: svn+ssh://multi2sim.org/multi2sim/trunk@816 a6503f54-2359-11de-8ce0-bfd7a313a9c3

+++ src/arch/evergreen/emu/isa.c
@@ -438,7 +438,7 @@ float evg_isa_read_op_src_float(int src_idx)
  * Deferred tasks for ALU group
  */
 
-void evg_isa_enqueue_write_lds(uint32_t addr, uint32_t value, size_t value_size)
+void evg_isa_enqueue_write_lds(uint32_t addr, uint32_t value, int value_size)
 {
 	struct evg_isa_write_task_t *wt;
